feat: Refactor event dispatcher and handlers directory structure
