+++ src/ai/backend/manager/event_dispatcher/handlers/BUILD
@@ -1,0 +1,1 @@
+python_sources(name="src")
