+++ src/ai/backend/manager/event_dispatcher/handlers/kernel.py
@@ -21,8 +21,8 @@ from ai.backend.common.types import (
 from ai.backend.logging import BraceStyleAdapter
 from ai.backend.manager.registry import AgentRegistry
 
-from ..models.kernel import kernels
-from ..models.utils import (
+from ...models.kernel import kernels
+from ...models.utils import (
     ExtendedAsyncSAEngine,
     execute_with_retry,
 )
